update to character.cs
A potential update to Character.cs

moved the character stats to private instead of public to protect them
and created classes to retrieve and change the stats as necessary.

Also added a class that is called when changes are done to update all
the character's derived stats.

Thinking of ways to rename stats to allow for more normalised naming
convention